+++ src/app/api/auth/[...nextauth]/route.ts
@@ -4,7 +4,6 @@ import NaverProvider from 'next-auth/providers/naver';
 import AppleProvider from 'next-auth/providers/apple';
 import NextAuth from 'next-auth';
 import { getAppleToken } from '@/utils/getAppleToken';
-import { SOCIAL_TYPE } from '@/types/Auth';
 import { AuthApi } from '../../AuthApi';
 
 const jwt = require('jsonwebtoken');
@@ -43,21 +42,11 @@ const handler = NextAuth({
   callbacks: {
     async signIn({ user, account }) {
       try {
-        let socialUniqueId = '';
-
-        if (account?.provider === 'apple' && account.id_token) {
-          const decoded = jwt.decode(account.id_token) as { sub: string };
-          socialUniqueId = decoded?.sub || '';
-        } else {
-          socialUniqueId = account?.providerAccountId || '';
-        }
-
         const body = {
           email: user.email as string,
           gender: null,
           age: null,
-          socialType: account?.provider as SOCIAL_TYPE,
-          socialUniqueId,
+          socialType: account?.provider as string,
         };
 
         const { accessToken, refreshToken } = await AuthApi.login(body);
