+++ src/app/api/oauth/kakao/route.ts
@@ -1,5 +1,4 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { SOCIAL_TYPE } from '@/types/Auth';
 import { AuthApi } from '../../AuthApi';
 
 const jwt = require('jsonwebtoken');
@@ -54,8 +53,7 @@ export async function POST(req: NextRequest) {
       email: userData.kakao_account?.email ?? 'no-email',
       gender: null,
       age: null,
-      socialType: SOCIAL_TYPE.KAKAO,
-      socialUniqueId: '',
+      socialType: 'KAKAO',
     };
 
     // Step 4: Log in with the Auth API
