+++ src/hooks/useWebViewInit.ts
@@ -5,39 +5,24 @@ import {
   handleWebViewMessage,
   sendLogToFlutter,
 } from '@/utils/flutterUtil';
-import { useAppSocialLogin } from './useAppSocialLogin';
 
 export const useWebViewInit = () => {
-  const {
-    onKakaoLoginSuccess,
-    onKakaoLoginError,
-    onAppleLoginSuccess,
-    onAppleLoginError,
-  } = useAppSocialLogin();
-  const [isMobile, setIsMobile] = useState<boolean | null>(null);
+  const [isMobile, setIsMobile] = useState(false);
 
   useEffect(() => {
-    if (typeof window === 'undefined') return;
-
     const { userAgent } = navigator;
     const mobile = /iPhone|iPad|iPod|Android/i.test(userAgent);
     setIsMobile(mobile);
   }, []);
 
   const initWebView = () => {
-    if (typeof window === 'undefined') return;
+    if (isMobile) {
+      handleWebViewMessage('checkIsInApp');
+    }
 
     window.getDeviceToken = getDeviceToken;
     window.checkIsInApp = checkIsInApp;
     window.sendLogToFlutter = sendLogToFlutter;
-    window.onKakaoLoginSuccess = onKakaoLoginSuccess;
-    window.onKakaoLoginError = onKakaoLoginError;
-    window.onAppleLoginSuccess = onAppleLoginSuccess;
-    window.onAppleLoginError = onAppleLoginError;
-
-    if (isMobile) {
-      handleWebViewMessage('checkIsInApp');
-    }
   };
 
   return { isMobile, initWebView };
