+++ src/utils/flutterUtil.ts
@@ -19,9 +19,7 @@ export function handleWebViewMessage(
     | 'deviceToken'
     | 'logToFlutter'
     | 'openAlbum'
-    | 'openCamera'
-    | 'loginWithKakao'
-    | 'loginWithApple',
+    | 'openCamera',
 ) {
   return window.FlutterMessageQueue.postMessage(message);
 }
@@ -32,9 +30,11 @@ export function openAlbum(imgDataBase64: string): string {
 
 export function sendLogToFlutter(log: string) {
   if (window.isInApp) {
+    // 웹뷰일 때 로그 출력
     window.LogToFlutter.postMessage(log);
     console.log(`[Message sent to Flutter - WebView] ${log}`);
   } else {
+    // 웹뷰가 아닐 때 로그 출력
     console.log(`[Message sent to Flutter - Browser] ${log}`);
   }
   return log;
